requirements.txt
torch>=2.2.0
transformers>=4.44.0
datasets>=2.20.0
accelerate>=0.30.0
peft>=0.11.0
pynvml>=11.5.0
mplfonts>=0.0.5
matplotlib>=3.8.0
pandas>=2.0.0
numpy>=1.26.0
pyyaml>=6.0
tqdm>=4.66.0
sentencepiece>=0.2.0
protobuf>=4.25.0
